==> 03 Scripts/Achievement 4/Step 2 IC Orders Products Combined.ipynb <==
# Step 2 Instacart Orders Products Combined

## Table of Contents
* [1. Introduction](#1.-Introduction)
* [2. Importing data](#2.-Importing-Data)
    * [2.1 Importing libraries](#2.1-Importing-Libraries)
    * [2.2 Importing data](#2.2-Importing-Data)
* [3. Data combining](#3.-Data-Combining)
    * [3.1 RAM memory space](#3.1-RAM-Memory-Space)
    * [3.2 Key variable](#3.2-Key-Variable)
    * [3.3 Merge](#3.3-Merge)
* [4. Exporting data](#4.-Exporting-Data) 

# 1. Introduction

Instacart's research question will be answered by merging every relevant dataframes. The Orders Products Prior dataframe and the Order dataframe will be integrated initially. By displaying all of the shopping items connected to each of the 3,421,083 total orders, this integration will improve each individual order.


# 2. Importing Data

## 2.1 Importing Libraries

import pandas as pd
import numpy as np
import os

## 2.2 Importing Data

path = r'C:\Users\kuohe\Instacart Basket Analysis 12-18-2023'

### Order

orders = pd.read_csv(os.path.join(path, '02 Data', 'Prepared data', 'orders_step1.csv'), index_col = False)

orders.shape

orders.head()

### Orders Products Prior

orders_products_prior = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'orders_products_prior_step1.pkl'))

orders_products_prior.shape

orders_products_prior.head()

# 3. Combining the Data

## 3.1 RAM Memory Space

### Orders

# Taking a look at the data types to save memory
orders.dtypes

# Converting the data type to be more efficient
orders['order_id'] = orders['order_id'].astype('int32')

# Converting the data type to be more efficient
orders['user_id'] = orders['user_id'].astype('int32')

# Converting the data type to be more efficient
orders['order_number'] = orders['order_number'].astype('int8')

# Converting the data type to be more efficient
orders['orders_day_of_week'] = orders['orders_day_of_week'].astype('int8')

# Converting the data type to be more efficient
orders['order_hour_of_day'] = orders['order_hour_of_day'].astype('int8')

# Checking if successful
orders.dtypes

### Orders Product Prior

orders_products_prior.dtypes

# Converting the data type to be more efficient
orders_products_prior['order_id'] = orders_products_prior['order_id'].astype('int32')

# Converting the data type to be more efficient
orders_products_prior['product_id'] = orders_products_prior['product_id'].astype('int32')

# Converting the data type to be more efficient
orders_products_prior['add_to_cart_sequence'] = orders_products_prior['add_to_cart_sequence'].astype('int32')

# Converting the data type to be more efficient
orders_products_prior['reordered'] = orders_products_prior['reordered'].astype('int8')

# Checking if successful
orders_products_prior.dtypes

## Key Variables

# Since order_id is shared among 2 dataframe, this will be the main focus for observation
orders['order_id'].describe()

unique_orders = orders['order_id'].nunique()
print(f"The amount of values in 'order_id': {unique_orders}")

# We will look at the order products prior for the "order_id" for observation
orders_products_prior['order_id'].describe()

unique_orders_products_prior = orders_products_prior['order_id'].nunique()
print(f"The amount of values in 'order_id': {unique_orders_products_prior}")

# 3.3 Merge

# Merging the dataframes
orders_products_combined = orders.merge(orders_products_prior, on = 'order_id')

# Checking if successful
orders_products_combined.head()

# Checking the amount of values in the combined order_id
unique_orders_products_combined = orders_products_combined['order_id'].nunique()
print(f"The amount of values in 'order_id': {unique_orders_products_combined}")

# 4. Exporting Data

orders_products_combined.shape

orders_products_combined.head()

# Exporting to prepared data folder
# This will be exported as a pickle due to the large size of data
orders_products_combined.to_pickle(os.path.join(path, '02 Data','Prepared data', 'orders_products_combined_step2.pkl'))
