==> 03 Scripts/Achievement 4/Step 3 IC Order Products Merged.ipynb <==
# Step 3 Order Products Merged

# Table of Contents
* [1. Introduction](#1.-Introduction)
* [2. Importing Data](#2.-Importing-Data)
    * [2.1 Importing Libraries](#2.1-Importing-Libraries)
    * [2.2 Importing Data](#2.2-Importing-Data)
* [3. Combining Data](#3.-Combining-Data)
    * [3.1 RAM Memory Space](#3.1-RAM-Memory-Space)
    * [3.2 Key Variables](#3.2-Key-Variables)
    * [3.3 Merge](#3.3-Merge)
* [4. Exporting Data](#4.-Exporting-Data) 

# 1. Introduction

Instacart's research question will be addressed by combining all dataframes in one comprehensive way. A unified dataframe called "orders_product_combined" was previously created by combining the Order and Orders Products Prior dataframes. Subsequently, this combined dataset will also incorporate the products dataframe. By doing this, the dataset will be enhanced with more product information, such as names and prices.


# 2. Importing Data

## 2.1 Importing Libraries

import pandas as pd
import numpy as np
import os

# 2.2 Importing Data

path = r'C:\Users\kuohe\Instacart Basket Analysis 12-18-2023'

### Order Products Combined

# Impiorting the file
orders_products_combined = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'orders_products_combined_step2.pkl'))

orders_products_combined.shape

orders_products_combined.head()

### Products

# Importing File
products = pd.read_csv(os.path.join(path, '02 Data', 'Prepared data', 'products_step1.csv'))

products.shape

products.head()

# 3. Cominbing Data

## 3.1 RAM Memory Space

### Order Products Combined

# Running some functions will take a lot of memory, so converting data types will help the process
orders_products_combined.dtypes

In step 2, some of the data types were already converted, so no furthur action needed

### Products

products.dtypes

# Converting the Data type to help the process
products['product_id'] = products['product_id'].astype('int32')

# Converting the Data type to help the process
products['aisle_id'] = products['aisle_id'].astype('int8')

# Converting the Data type to help the process
products['department_id'] = products['department_id'].astype('int8')

# Check if implemented successfully
products.dtypes

# 3.2 Key Variables

# "product_id" is a shared variable among 2 dataframes so this will be the main focus
orders_products_combined['product_id'].describe()

# Double checking the values in "product_id" from the orders product combined file
values_orders_products_combined = orders_products_combined['product_id'].nunique()
print(f"The amoount of values in 'product_id': {values_orders_products_combined}")

# Now checking it from the products file for "product_id"
products['product_id'].describe()

values_products = products['product_id'].nunique()
print(f"The amounts of values in 'product_id': {values_products}")

# 3.3 Merge

# Merging the dataframes
orders_products_merged = orders_products_combined.merge(products, on = 'product_id')

# Checking
orders_products_merged.head()

# 4. Exporting Data

# Check size
orders_products_merged.shape

orders_products_merged.head()

# Exporting to prepared data folder
# The file type should be pickle sinze there are over 32 million rows
orders_products_merged.to_pickle(os.path.join(path, '02 Data','Prepared data', 'orders_products_merged_step3.pkl'))
