==> 03 Scripts/Achievement 4/Step 4 IC Orders Products All.ipynb <==
# Step 4 Instacart Orders Products All

# Table of Contents
* [1. Introduction](#1.-Introduction)
* [2. Importing Data](#2.-Importing-Data)
    * [2.1 Importing Libraries](#2.1-Importing-Libraries)
    * [2.2 Importing Data](#2.2-Importing-Data)
* [3. Combining Data](#3.-Combining-Data)
    * [3.1 RAM Memory Space](#3.1-RAM-Memory-Space)
    * [3.2 Key Variable](#3.2-Key-Variable)
    * [3.3 Merge](#3.3-Merge)
* [4. Exporting Data](#4.-Exporting-Data) 

# 1. Introduction

A thorough connection of all relevant dataframes is in progress to address Instacart's research question. The Order and Orders Products Prior dataframes were first combined to form the "orders_product_combined" dataframe. The products dataframe was then added as well, creating the new dataframe "orders_product_merged". The Customer dataframe must be added to this amalgamation in the following step. This enhancement will make it possible to link orders to particular customers, making it easier to analyze consumer purchasing patterns in-depth.


# 2. Importing Data

## 2.1 Importing Libraries

import pandas as pd
import numpy as np
import os

## 2.2 Importing Data

path = r'C:\Users\kuohe\Instacart Basket Analysis 12-18-2023'

orders_products_merged = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'orders_products_merged_step3.pkl'))

# Checking Size
orders_products_merged.shape

# Checking Columns
orders_products_merged.head()

### Customers

customers = pd.read_csv(os.path.join(path, '02 Data', 'Prepared data', 'customers_step1.csv'))

# Check Size
customers.shape

#Check Columns
customers.head()

# 3. Combining Data

## 3.1 RAM Memory Space

### Orders Product Merged

# Checking data type to save RAM
orders_products_merged.dtypes

Most of the data types have already been converted in previous steps

### Customers

# Checking data type to save RAM
customers.dtypes

# Converting the data type to improve processing speed
customers['user_id'] = customers['user_id'].astype('int32')

# Converting the data type to improve processing speed
customers['Gender'] = customers['Gender'].astype('category')

# Converting the data type to improve processing speed
customers['STATE'] = customers['STATE'].astype('category')

# Converting the data type to improve processing speed
customers['Age'] = customers['Age'].astype('int32')

# Converting the data type to improve processing speed
customers['No._of_dependants'] = customers['No._of_dependants'].astype('int32')

# Converting the data type to improve processing speed
customers['family_status'] = customers['family_status'].astype('category')

# 3.2 Key Variables

# Since "user_id is shared among 2 dataframes, it will be the main focus"
orders_products_merged['user_id'].describe()

values_orders_products_merged = orders_products_merged['user_id'].nunique()
print(f"The amount of values in 'user_id': {values_orders_products_merged}")

# As "user_id" is also in the customer file, we must check the values as well
customers['user_id'].describe()

values_customers = customers['user_id'].nunique()
print(f"The amount of values in 'user_id': {values_customers}")

# 3.3 Merge

# Merging the dataframes
orders_products_all = orders_products_merged.merge(customers, on = 'user_id')

# Checking Columns
orders_products_all.head()

# Double checking the values if the merge was successful, it should be the same values
values_orders_products_all = orders_products_all['user_id'].nunique()
print(f"The amount of values in 'user_id': {values_orders_products_all}")

orders_products_all['user_id'].describe()

It was merged successfully

# 4. Exporting Data

# Checking Size
orders_products_all.shape

# Checking Column name
orders_products_all.head()

# Exporting to prepared data folder
# Since there is over 32 million rows of data this must be exported as a pickle  
orders_products_all.to_pickle(os.path.join(path, '02 Data','Prepared data', 'orders_products_all_step4.pkl'))
