==> 03 Scripts/Achievement 4/Step 5 IC Exclusion Flag for Customers.ipynb <==
# Step 5 Instacart Exclusion Flag for Customers

# Table of Contents
* [1. Introduction](#1.-Introduction)
* [2. Importing Data](#2.-Importing-Data)
    * [2.1 Importing Libraries](#2.1-Importing-Libraries)
    * [2.2 Importing Data](#2.2-Importing-Data)
* [3. Exclusion Flag](#3.-Exclusion-Flag)
    * [3.1 Aggregate Variables](#3.1-Aggregate-Variables)
    * [3.2 Derived Variables](#3.2-Derived-Variables)
    * [3.3 Final Dataframe](#3.3-Final-Dataframe)
* [4. Exporting Data](#4.-Exporting-Data) 

# 1. Introduction

All of the pertinent dataframes have been combined into a single, extensive dataframe in order to answer Instacart's research question. We'll use a filtering approach because the Instacart CFO is focused on customers who make a substantial contribution to the app's revenue. Customers with less than five orders, or those with low activity, will be eliminated from the 'orders_products_all' dataframe. The goal of this step is to produce a final dataframe for more focused analysis that represents the behavior of higher-value clients.


# 2. Importing Data

## 2.1 Importing Libraries

import pandas as pd
import numpy as np
import os

## 2.2 Importing Data

path = r'C:\Users\kuohe\Instacart Basket Analysis 12-18-2023'

orders_products_all = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'orders_products_all_step4.pkl'))

orders_products_all.shape

orders_products_all.head()

# 3. Exclusion Flag

## 3.1 Aggregate Variables

# Observing what is the max number of order that each user have ordered
orders_products_all['max_order_number'] = orders_products_all.groupby(['user_id'])['order_number'].transform(np.max)

# Checking results
orders_products_all.head()

# Checking aggregations
orders_products_all.groupby(['user_id'])['order_number'].max()

# Creating a dataframe for checking the max order number
df = orders_products_all[['user_id', 'max_order_number']]

# Subgroupping user_id to check
df.loc[df['user_id'].isin([5, 206208])]

# 3.2 Derived Variable

# Creating the exclusion for low activity customers
# This code is for customers that have 5 orders or less
orders_products_all.loc[orders_products_all['max_order_number'] < 5, 'activity_customer'] = 'low activity customer'

# This code is for customers that have 5 orders or more
orders_products_all.loc[orders_products_all['max_order_number'] >= 5, 'activity_customer'] = 'regular customer'

# Checking if applied
orders_products_all.groupby(['user_id'])['activity_customer'].max()

# Creating another dataframe for checking activity of customer
df = orders_products_all[['user_id', 'max_order_number', 'activity_customer']] 

# Subgroupping user_id to check 
df.loc[df['user_id'].isin([5, 206208])]

# Checking frequency of flag of "activity_customer"
orders_products_all['activity_customer'].value_counts(dropna = False) 

# 3.3 Final Dataframe

# Creating exclusion flag for low activity customers
orders_products_final = orders_products_all.loc[orders_products_all['activity_customer'] == 'regular customer' ]

orders_products_final.head()

# Checking the frequency for "activity_customer"
orders_products_final['activity_customer'].value_counts(dropna = False)

orders_products_final.shape

# Since the data is filtered, we can drop the column to show only the regular customer since the CFO does not want to see low activity customers
orders_products_final = orders_products_final.drop(columns = ['activity_customer'])

# Check column 
orders_products_final.head()

# 4. Exporting Data

# Exporting to prepared data folder
# Since there are 31 million rows of data it is better to export in pickle format for easier access
orders_products_final.to_pickle(os.path.join(path, '02 Data','Prepared data', 'orders_products_final_step5.pkl'))
